福利接口代码

+++ app/src/main/java/com/qicheng/business/module/Benefit.java
@@ -28,13 +28,13 @@ public class Benefit implements Serializable {
      * 1:已使用/已领取
      * 2:已过期
      */
-    private byte status;
+    private int status;
     /**
      * 后续可操作标识
      * 0:不可操作
      * 1:可转让
      */
-    private byte postOpFlag;
+    private int postOpFlag;
 
     /**
      * 当前来源ID
@@ -68,7 +68,7 @@ public class Benefit implements Serializable {
     /**
      * 福利金额
      */
-    private int value;
+    private double value;
     /**
      * 福利说明
      */
@@ -94,19 +94,19 @@ public class Benefit implements Serializable {
         this.name = name;
     }
 
-    public byte getStatus() {
+    public int getStatus() {
         return status;
     }
 
-    public void setStatus(byte status) {
+    public void setStatus(int status) {
         this.status = status;
     }
 
-    public byte getPostOpFlag() {
+    public int getPostOpFlag() {
         return postOpFlag;
     }
 
-    public void setPostOpFlag(byte postOpFlag) {
+    public void setPostOpFlag(int postOpFlag) {
         this.postOpFlag = postOpFlag;
     }
 
@@ -166,11 +166,11 @@ public class Benefit implements Serializable {
         this.expireTime = expireTime;
     }
 
-    public int getValue() {
+    public double getValue() {
         return value;
     }
 
-    public void setValue(int value) {
+    public void setValue(double value) {
         this.value = value;
     }
